+++ src/igate.go
@@ -1086,7 +1086,7 @@ func igate_recv_thread() {
 					var slice C.int = 0
 					var fec_type fec_type_t = fec_type_none
 					var spectrum = C.CString("APRS-IS")
-					dlq_rec_frame(ichan, subchan, slice, pp3, alevel, fec_type, RETRY_NONE, spectrum)
+					dlq_rec_frame(C.int(ichan), subchan, slice, pp3, alevel, fec_type, RETRY_NONE, spectrum)
 				} else {
 					text_color_set(DW_COLOR_ERROR)
 					dw_printf("ICHANNEL %d: Could not parse message from APRS-IS server.\n", ichan)
